Refactor: refine use of path_closure fn
Before it was necessary to call path_closure and assign to path_self in
the global environment to run the model. Now path_closure is assigned to
path_self within the higher level fns reusing the arguments from the
higher level fn - and if required passed as an argument to lower level
fns. All higher level fns are now called with as a minimum area_code and
base_year parameters.