Fix: prevent binning fn failing for small ranges
For small ranges `nclass.FD` will suggest a large number of classes
(eg >100) - I'm not sure this is desirable. In addition, when `summary`
is used on the intervals generated by `cut`, by default it only handles
100 levels for factors (anything more is captured as "Other"). If this
happens the binning falls over as the breaks and intervals don't line
up. Fix automatically suggests 20 classes if the range is <.1 otherwise
`nclass.FD` is used, for extra insurance maxsum argument of `summary` is
set to 200.